added confirmation dialog for deleting a blog post

+++ part5/bloglist-frontend/src/components/Blog.js
@@ -42,8 +42,10 @@ const Blog = ({blog, updateBlog, deleteBlog, loggedUser}) => {
     e.preventDefault()
     const elId = newBlog.id
 
-    console.log('delete this blog', elId,newBlog)
-    deleteBlog(elId, newBlog)
+    
+    if(window.confirm(`Remove blog ${newBlog.title} by ${newBlog.author}`)) {
+      deleteBlog(elId, newBlog)
+    }
   }
